Removing Duplicated before return
